requirements.txt
websockets>=12.0
aiohttp>=3.9.3
sortedcontainers>=2.4.0
pandas>=2.2.1
numpy>=1.26.4
python-telegram-bot>=21.3
python-dotenv>=1.0.1
pyyaml>=6.0.1
structlog>=24.1.0

# Autoresearch optimization loop
openai>=1.30.0

# Optional: on-chain logging
# web3>=6.15.1

# Optional: PostgreSQL persistence
# psycopg2-binary>=2.9.9

# Optional: Redis caching
# redis>=5.0.3
